+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractValue.java
@@ -351,7 +351,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineNoAnnotations(
+    protected @Nullable AnnotationMirror combineTwoTypeVars(
         AnnotatedTypeVariable aAtv,
         AnnotatedTypeVariable bAtv,
         AnnotationMirror top,
@@ -366,7 +366,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineOneAnnotation(
+    protected @Nullable AnnotationMirror combineAnnotationWithTypeVar(
         AnnotationMirror annotation,
         AnnotatedTypeVariable typeVar,
         AnnotationMirror top,
@@ -483,7 +483,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineNoAnnotations(
+    protected @Nullable AnnotationMirror combineTwoTypeVars(
         AnnotatedTypeVariable aAtv,
         AnnotatedTypeVariable bAtv,
         AnnotationMirror top,
@@ -499,7 +499,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineOneAnnotation(
+    protected @Nullable AnnotationMirror combineAnnotationWithTypeVar(
         AnnotationMirror annotation,
         AnnotatedTypeVariable typeVar,
         AnnotationMirror top,
@@ -581,7 +581,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineNoAnnotations(
+    protected @Nullable AnnotationMirror combineTwoTypeVars(
         AnnotatedTypeVariable aAtv,
         AnnotatedTypeVariable bAtv,
         AnnotationMirror top,
@@ -597,7 +597,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
     }
 
     @Override
-    protected @Nullable AnnotationMirror combineOneAnnotation(
+    protected @Nullable AnnotationMirror combineAnnotationWithTypeVar(
         AnnotationMirror annotation,
         AnnotatedTypeVariable typeVar,
         AnnotationMirror top,
@@ -634,8 +634,8 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
    *
    * <ol>
    *   <li>{@link #combineTwoAnnotations}
-   *   <li>{@link #combineOneAnnotation}
-   *   <li>{@link #combineNoAnnotations}
+   *   <li>{@link #combineAnnotationWithTypeVar}
+   *   <li>{@link #combineTwoTypeVars}
    * </ol>
    *
    * If a set is missing an annotation in a hierarchy, and if the combined set can be missing an
@@ -679,11 +679,11 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
         if (a != null && b != null) {
           result = combineTwoAnnotations(a, b, top);
         } else if (a != null) {
-          result = combineOneAnnotation(a, bAtv, top, canCombinedSetBeMissingAnnos);
+          result = combineAnnotationWithTypeVar(a, bAtv, top, canCombinedSetBeMissingAnnos);
         } else if (b != null) {
-          result = combineOneAnnotation(b, aAtv, top, canCombinedSetBeMissingAnnos);
+          result = combineAnnotationWithTypeVar(b, aAtv, top, canCombinedSetBeMissingAnnos);
         } else {
-          result = combineNoAnnotations(aAtv, bAtv, top, canCombinedSetBeMissingAnnos);
+          result = combineTwoTypeVars(aAtv, bAtv, top, canCombinedSetBeMissingAnnos);
         }
         if (result != null) {
           combinedSets.add(result);
@@ -716,7 +716,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
      * @param canCombinedSetBeMissingAnnos whether or not
      * @return the result of combining the two type variables, which may be null
      */
-    protected abstract @Nullable AnnotationMirror combineNoAnnotations(
+    protected abstract @Nullable AnnotationMirror combineTwoTypeVars(
         AnnotatedTypeVariable aAtv,
         AnnotatedTypeVariable bAtv,
         AnnotationMirror top,
@@ -733,7 +733,7 @@ public abstract class CFAbstractValue<V extends CFAbstractValue<V>> implements A
      * @param canCombinedSetBeMissingAnnos whether or not
      * @return the result of combining {@code annotation} with {@code typeVar}
      */
-    protected abstract @Nullable AnnotationMirror combineOneAnnotation(
+    protected abstract @Nullable AnnotationMirror combineAnnotationWithTypeVar(
         AnnotationMirror annotation,
         AnnotatedTypeVariable typeVar,
         AnnotationMirror top,